New electrical things

+++ src/main/java/com/alpha67/amc/procedures/A0Procedure.java
@@ -42,7 +42,7 @@ public class A0Procedure {
 					return tileEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "timerGui") == 0) {
+		}.getValue(world, new BlockPos(x, y, z), "timerGui") == 0) {
 			return true;
 		}
 		return false;
