+++ src/main/java/com/alpha67/amc/procedures/A2Procedure.java
@@ -42,7 +42,7 @@ public class A2Procedure {
 					return tileEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "timerGui") == 2) {
+		}.getValue(world, new BlockPos(x, y, z), "timerGui") == 2) {
 			return true;
 		}
 		return false;
