+++ src/main/java/com/alpha67/amc/procedures/A5Procedure.java
@@ -42,7 +42,7 @@ public class A5Procedure {
 					return tileEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "timerGui") == 5) {
+		}.getValue(world, new BlockPos(x, y, z), "timerGui") == 5) {
 			return true;
 		}
 		return false;
