+++ src/main/java/com/alpha67/amc/procedures/Lavabuketprocedure5Procedure.java
@@ -42,7 +42,7 @@ public class Lavabuketprocedure5Procedure {
 					return tileEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "bucketoflava") == 5) {
+		}.getValue(world, new BlockPos(x, y, z), "bucketoflava") == 5) {
 			return true;
 		}
 		return false;
