+++ src/main/java/com/alpha67/amc/procedures/Lavabuketprocedure7Procedure.java
@@ -42,7 +42,7 @@ public class Lavabuketprocedure7Procedure {
 					return tileEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "bucketoflava") == 7) {
+		}.getValue(world, new BlockPos(x, y, z), "bucketoflava") == 7) {
 			return true;
 		}
 		return false;
